Merge: Fix security issues -> develop

+++ src/forms/subscription-delete.ts
@@ -1,47 +1,49 @@
-import { AppCallRequest, AppCallValues, AppForm, AppFormValue, AppSelectOption, Subscription } from '../types';
-import { ConfigStoreProps, KVStoreClient, KVStoreOptions } from '../clients/kvstore';
-import { AppExpandLevels, AppFieldTypes, ExceptionType, OpsGenieIcon, Routes, StoreKeys, SubscriptionCreateForm, SubscriptionDeleteForm } from '../constant';
+import { AppActingUser, AppCallRequest, AppCallValues, AppForm, AppSelectOption, Integration, Subscription, Teams } from '../types';
+import { AppExpandLevels, AppFieldTypes, ExceptionType, OpsGenieIcon, Routes, SubscriptionDeleteForm } from '../constant';
 import { OpsGenieClient, OpsGenieOptions } from '../clients/opsgenie';
 import { configureI18n } from '../utils/translations';
-import { getIntegrationsList, tryPromise } from '../utils/utils';
+import { getIntegrationsList, isUserSystemAdmin, tryPromise } from '../utils/utils';
 import { Exception } from '../utils/exception';
+import { ExtendRequired, allowMemberAction, getOpsGenieAPIKey } from '../utils/user-mapping';
+
+import { getAllTeamsCall } from './list-team';
 
 export async function subscriptionDeleteCall(call: AppCallRequest): Promise<string> {
-    const mattermostUrl: string = call.context.mattermost_site_url!;
-    const botAccessToken: string = call.context.bot_access_token!;
+    const actingUser: AppActingUser | undefined = call.context.acting_user;
+    const isSystemAdmin: boolean = isUserSystemAdmin(actingUser);
+    const allowMember: boolean = allowMemberAction(call.context);
     const values: AppCallValues | undefined = call.values;
     const i18nObj = configureI18n(call.context);
-
-    const subscription: AppSelectOption = values?.[SubscriptionDeleteForm.SUBSCRIPTION_ID];
-
-    const options: KVStoreOptions = {
-        mattermostUrl,
-        accessToken: botAccessToken,
-    };
-    const kvStore: KVStoreClient = new KVStoreClient(options);
-
-    const configStore: ConfigStoreProps = await kvStore.kvGet(StoreKeys.config);
+    const apiKey = getOpsGenieAPIKey(call);
 
     const optionsOps: OpsGenieOptions = {
-        api_key: configStore.opsgenie_apikey,
+        api_key: apiKey,
     };
     const opsGenieClient: OpsGenieClient = new OpsGenieClient(optionsOps);
+    const subscription: AppSelectOption = values?.[SubscriptionDeleteForm.SUBSCRIPTION_ID];
+
+    if (!allowMember) {
+        throw new Exception(ExceptionType.MARKDOWN, i18nObj.__('general.validation-user.genie-action-invalid'));
+    }
+
+    if (!isSystemAdmin) {
+        const opsSubscription: Integration = await tryPromise<Integration>(opsGenieClient.getIntegration(subscription.value), ExceptionType.MARKDOWN, i18nObj.__('forms.error'));
+
+        const teams: Teams[] = await getAllTeamsCall(call);
+        const teamsIds: string[] = teams.map((team) => team.id);
+
+        if (!teamsIds.includes(opsSubscription.ownerTeam.id)) {
+            throw new Exception(ExceptionType.MARKDOWN, i18nObj.__('binding.binding.command-delete-no-found'));
+        }
+    }
 
     await tryPromise(opsGenieClient.deleteIntegration(subscription.value), ExceptionType.MARKDOWN, i18nObj.__('forms.error'));
     return i18nObj.__('api.subcription.message-delete');
 }
 
 export async function subscriptionDeleteFormCall(call: AppCallRequest): Promise<AppForm> {
-    const mattermostUrl: string | undefined = call.context.mattermost_site_url;
-    const botAccessToken: string | undefined = call.context.bot_access_token;
     const i18nObj = configureI18n(call.context);
-
-    const options: KVStoreOptions = {
-        mattermostUrl: <string>mattermostUrl,
-        accessToken: <string>botAccessToken,
-    };
-
-    const integrations: Subscription[] = await getIntegrationsList(options, i18nObj);
+    const integrations: Subscription[] = await getIntegrationsList(call);
 
     if (!integrations.length) {
         throw new Exception(ExceptionType.MARKDOWN, i18nObj.__('binding.binding.command-delete-no-subscriptions'));
@@ -60,9 +62,8 @@ export async function subscriptionDeleteFormCall(call: AppCallRequest): Promise<
         submit: {
             path: Routes.App.CallPathSubscriptionDeleteSubmit,
             expand: {
+                ...ExtendRequired,
                 app: AppExpandLevels.EXPAND_ALL,
-                oauth2_app: AppExpandLevels.EXPAND_ALL,
-                oauth2_user: AppExpandLevels.EXPAND_ALL,
             },
         },
         fields: [
